requirements.txt
absl-py==2.0.0
aiohttp==3.8.6
aiosignal==1.3.1
alabaster==0.7.13
aniso8601==9.0.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asciitree==0.3.3
asttokens==2.4.0
async-timeout==4.0.3
attrdict==2.0.1
attrs==23.1.0
audioread==3.0.1
Babel==2.13.0
backcall==0.2.0
beautifulsoup4==4.12.2
black==19.10b0
blis==0.7.11
boto3==1.28.63
botocore==1.31.63
braceexpand==0.1.7
cachetools==5.3.1
catalogue==2.0.10
cdifflib==1.2.6
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.0.2
cloudpathlib==0.15.1
colorama==0.4.6
comm==0.1.4
confection==0.1.3
contourpy==1.1.1
cycler==0.12.1
cymem==2.0.8
Cython==3.0.3
datasets==2.14.5
decorator==5.1.1
dill==0.3.7
Distance==0.1.3
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.20.1
editdistance==0.6.2
einops==0.7.0
exceptiongroup==1.1.3
executing==2.0.0
faiss-cpu==1.7.4
fasteners==0.19
fasttext==0.9.2
filelock==3.12.4
Flask==2.2.5
Flask-RESTful==0.3.10
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.1
g2p-en==2.1.0
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.37
google-auth==2.23.3
google-auth-oauthlib==1.0.0
grpcio==1.59.0
h5py==3.10.0
huggingface-hub==0.18.0
hydra-core==1.3.2
idna==3.4
ijson==3.2.3
imagesize==1.4.1
inflect==7.0.0
iniconfig==2.0.0
ipython==8.16.1
ipywidgets==8.1.1
isort==5.12.0
itsdangerous==2.1.2
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.2
jiwer==2.5.2
jmespath==1.0.1
joblib==1.3.2
jupyterlab-widgets==3.0.9
kaldi-python-io==1.2.2
kaldiio==2.18.0
kiwisolver==1.4.5
kornia==0.7.0
langcodes==3.3.0
latexcodec==2.0.1
lazy_loader==0.3
Levenshtein==0.22.0
librosa==0.10.1
lightning-utilities==0.9.0
llvmlite==0.41.0
loguru==0.7.2
lxml==4.9.3
Markdown==3.5
markdown-it-py==3.0.0
markdown2==2.4.10
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
megatron-core==0.2.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.10
nemo-text-processing==0.2.0rc0
networkx==3.1
nltk==3.8.1
numba==0.58.0
numcodecs==0.12.0
numpy==1.23.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.2.140
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.14.1
OpenCC==1.1.6
packaging==23.2
pandas==2.1.1
pangu==4.0.6.1
parameterized==0.9.0
parso==0.8.3
pathspec==0.11.2
pathtools==0.1.2
pathy==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
plac==1.4.0
platformdirs==3.11.0
pluggy==1.3.0
pooch==1.7.0
portalocker==2.8.2
preshed==3.0.9
progress==1.6
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pybind11==2.11.1
pybtex==0.24.0
pybtex-docutils==1.0.3
pycparser==2.21
pydantic==1.10.13
pydub==0.25.1
Pygments==2.16.1
pynini==2.1.5
pyparsing==3.1.1
pypinyin==0.49.0
pypinyin-dict==0.6.0
PySocks==1.7.1
pytest==7.4.2
pytest-runner==6.0.0
python-dateutil==2.8.2
pytorch-lightning==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==2.13.7
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
rouge-score==0.1.2
rsa==4.9
ruamel.yaml==0.17.35
ruamel.yaml.clib==0.2.8
s3transfer==0.7.0
sacrebleu==2.3.1
sacremoses==0.0.53
safetensors==0.4.0
scikit-learn==1.3.1
scipy==1.11.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.32.0
setproctitle==1.3.3
shellingham==1.5.3
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
sox==1.4.1
soxr==0.3.7
spacy==3.7.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-bibtex==2.6.1
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
srsly==2.4.8
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorstore==0.1.45
termcolor==2.3.0
text-unidecode==1.3
textdistance==4.6.0
texterrors==0.4.4
thinc==8.2.1
threadpoolctl==3.2.0
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
torch==2.1.0
torchmetrics==1.2.0
torchvision==0.16.0
tqdm==4.66.1
traitlets==5.11.2
transformers==4.33.3
triton==2.1.0
typed-ast==1.5.5
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
wandb==0.15.12
wasabi==1.1.2
wcwidth==0.2.8
weasel==0.3.2
webdataset==0.1.62
Werkzeug==3.0.0
wget==3.2
widgetsnbextension==4.0.9
wrapt==1.15.0
xxhash==3.4.1
yarl==1.9.2
youtokentome==1.0.6
zarr==2.16.1
